train/requirements.txt
torch==2.1.2
transformers==4.36.2
datasets==2.16.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.3
spacy==3.7.2
nltk==3.8.1
pymongo==4.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
accelerate==0.25.0